backend/service/scraping-service/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
apify-client>=1.3.0